renamed qaos in quaos
