+++ .github/CODEOWNERS
@@ -1,0 +1,3 @@
+* @c-nation @f-martini @ldellant @RickSimon79221
+
+/quaos/ @c-nation @f-martini @ldellant @RickSimon79221 @cerisola @AndrewJena
