--- docs/index/_autosummary/qaos.pauli.rst
@@ -1,12 +1,0 @@
-qaos.pauli
-==========
-
-.. automodule:: qaos.pauli
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      pauli
-
