+++ docs/index/_autosummary/quaos.pauli.rst
@@ -1,0 +1,24 @@
+quaos.pauli
+===========
+
+.. automodule:: quaos.pauli
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      pauli_product
+      pauli_to_matrix
+      pauli_to_string
+      quditwise_inner_product
+      string_to_pauli
+      symplectic_inner_product
+      tensor
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      pauli
+
